add jsonnet templates for deployment
This change adds jsonnet templates for configuring cortex. These
templates are intended to be extensible and easily configurable for
vendoring into a repo and layering any tweaks on top.

There are two entrypoints to the templates
- cortex.jsonnet - provides all cortex components
- demo.jsonnet - provides all cortex deployments and the basic
configuration of prometheus & nginx for testing and demo purposes

The demo includes a prometheus which is configured to scrape all cortex
components and leverage remote storage into the cortex deployment.

Signed-off-by: Graham Rounds <graham@platform9.com>